Декомпозиция редьюсера, создание селекторов, подключение в app
