+++ src/components/movie-page/movie-page.jsx
@@ -10,7 +10,13 @@ const MoviesListWrapped = withActiveItem(MoviesList);
 const TabsWrapped = withActiveItem(Tabs);
 const MoviePlayerWrapped = withVideo(MoviePlayer);
 
+const getSameGenreFilms = (films, activeMovie) => {
+  return films.filter((film) => film.genre === activeMovie.genre).filter((film) => film.id !== activeMovie.id).slice(0, 4);
+};
+
 const MoviePage = ({movie, films, onMovieCardClick, playingFilm, setPlayingFilm}) => {
+  const filmsToRender = getSameGenreFilms(films, movie);
+
   return playingFilm
     ? (<MoviePlayerWrapped movie={playingFilm} onExitClick={()=> setPlayingFilm(null)}/>)
     : (<React.Fragment>
@@ -82,7 +88,7 @@ const MoviePage = ({movie, films, onMovieCardClick, playingFilm, setPlayingFilm}
         <section className="catalog catalog--like-this">
           <h2 className="catalog__title">More like this</h2>
 
-          <MoviesListWrapped films={films} onMovieCardClick={onMovieCardClick}/>
+          <MoviesListWrapped films={filmsToRender} onMovieCardClick={onMovieCardClick}/>
         </section>
 
         <footer className="page-footer">
