+++ src/reducer/state/selectors.js
@@ -1,0 +1,20 @@
+import NameSpace from "../name-space.js";
+
+
+const NAME_SPACE = NameSpace.STATE;
+
+export const getActiveFilm = (state) => {
+  return state[NAME_SPACE].activeFilm;
+};
+
+export const getPlayingFilm = (state) => {
+  return state[NAME_SPACE].playingFilm;
+};
+
+export const getShownCardsBound = (state) => {
+  return state[NAME_SPACE].shownCardsBound;
+};
+
+export const getActiveGenre = (state) => {
+  return state[NAME_SPACE].activeGenre;
+};
